Update form validator to be able to check email and submit are present columns
[MAILPOET-2754]
